Python/PyMail/requirements.txt
click==8.1.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.20.1
tomli==2.0.1
